pkg/stats: split out pkg/stats/sample
This will reduce the number of dependencies needed for the main
syzkaller tools.
