+++ pkg/stats/sample/sample_test.go
@@ -1,7 +1,7 @@
 // Copyright 2021 syzkaller project authors. All rights reserved.
 // Use of this source code is governed by Apache 2 LICENSE that can be found in the LICENSE file.
 
-package stats
+package sample
 
 import (
 	"reflect"
